SnowGroundArray: Use int32_t instead of long for members

+++ include/OSNOWG.h
@@ -25,6 +25,8 @@
 #ifndef __SNOWG_H
 #define __SNOWG_H
 
+#include <stdint.h>
+
 #include <ODYNARRB.h>
 
 class SnowGround
@@ -63,8 +65,8 @@ private:
 	unsigned	seed;
 
 public:
-	long	snow_thick;
-	long	snow_pattern;
+	int32_t snow_thick;
+	int32_t snow_pattern;
 
 public:
 	void	init(long initSnowScale, long anyNumber);
